# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Music Album API
  description: API for creating and managing music albums
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Albums
    description: Album management operations

paths:
  /api/v1/albums:
    post:
      tags:
        - Albums
      operationId: createAlbum
      summary: Creates an album at NEW status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumRequest'
      responses:
        '201':
          description: Album created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumServiceError'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumServiceError'

    get:
      tags:
        - Albums
      operationId: getAllAlbums
      summary: List all albums
      responses:
        '200':
          description: List of all albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlbumResponse'

  /api/v1/albums/{id}:
    get:
      tags:
        - Albums
      operationId: getAlbumById
      summary: Get album by ID
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumServiceError'

    put:
      tags:
        - Albums
      operationId: updateAlbum
      summary: Update album properties
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumRequest'
      responses:
        '200':
          description: Album updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumServiceError'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumServiceError'

    delete:
      tags:
        - Albums
      operationId: deleteAlbum
      summary: Delete album
      description: Marks the album as DELETED. Only allowed when the album is in PUBLISHED status.
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '204':
          description: Album deleted
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumServiceError'
        '422':
          description: Album is not in PUBLISHED status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumServiceError'

components:
  parameters:
    AlbumId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    CreateTrackRequest:
      type: object
      required:
        - name
        - src
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        src:
          type: string
          minLength: 1

    CreateAlbumRequest:
      type: object
      required:
        - name
        - imageSrc
        - artistId
        - tracks
        - destinations
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        imageSrc:
          type: string
          minLength: 1
        artistId:
          type: integer
          format: int64
        tracks:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateTrackRequest'
        destinations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AlbumDestinations'

    UpdateAlbumRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    AlbumStatusEnum:
      type: string
      enum:
        - NEW
        - VALIDATED
        - PUBLISHED
        - DELETED
    AlbumDestinations:
      type: string
      enum:
        - SPOTIFY
    ArtistResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        albumCount:
          type: integer

    TrackResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        src:
          type: string

    AlbumResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/AlbumStatusEnum'
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/AlbumDestinations'
        imageSrc:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistResponse'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/TrackResponse'

    AlbumServiceError:
      type: object
      properties:
        errorCode:
          type: string
        messages:
          type: array
          items:
            type: string